added module and callbacks

+++ tul_net_context.h
@@ -13,8 +13,8 @@
 typedef struct __tul_net_context
 {
   unsigned _sock;
-  unsigned long long _tsend;
-  unsigned long long _trecv;
+  unsigned _tsend;
+  unsigned _trecv;
   unsigned payload_out_cnt;
   unsigned char payload_in[CTX_BLOCK];
   unsigned char payload_out[CTX_BLOCK];
